--- src/main/java/ozil/Command.java
@@ -1,10 +1,0 @@
-package ozil;
-
-public class Command {
-    public void run(TaskList tasks) throws OzilException{
-    }
-
-    public boolean isTerminatingCommand() {
-        return false;
-    }
-}
