+++ src/main/java/ozil/command/AddDeadlineTaskCommand.java
@@ -1,12 +1,16 @@
-package ozil;
+package ozil.command;
 
+import ozil.exception.ErrorMessages;
+import ozil.main.Messages;
+import ozil.exception.OzilException;
+import ozil.main.TaskList;
 import ozil.task.DeadlineTask;
 
-public class AddDeadlineTaskCommand extends Command{
+public class AddDeadlineTaskCommand extends Command {
     private String description;
     private String deadline;
 
-    public AddDeadlineTaskCommand(String userInput) throws OzilException{
+    public AddDeadlineTaskCommand(String userInput) throws OzilException {
         String[] sections = userInput.split("\\s+", 2);
 
         if (sections[1].trim().startsWith("/by")) {
