+++ src/main/java/ozil/command/AddTodoTaskCommand.java
@@ -1,11 +1,15 @@
-package ozil;
+package ozil.command;
 
+import ozil.exception.ErrorMessages;
+import ozil.main.Messages;
+import ozil.exception.OzilException;
+import ozil.main.TaskList;
 import ozil.task.TodoTask;
 
-public class AddTodoTaskCommand extends Command{
+public class AddTodoTaskCommand extends Command {
     private String userInputDescription;
 
-    public AddTodoTaskCommand(String userInput) throws OzilException{
+    public AddTodoTaskCommand(String userInput) throws OzilException {
         String[] sections = userInput.split("\\s+", 2);
         if (sections.length < 2) {
             throw new OzilException(ErrorMessages.taskDescriptionError("todo"));
