+++ src/main/java/ozil/main/Storage.java
@@ -1,4 +1,4 @@
-package ozil;
+package ozil.main;
 
 public class Storage {
